fix+feat: optimize and fix log viewer's selection
This both fixes an issue created in the previous commit with selection
inside the line (as oposed to multi-line selections), and optimizes it
to not freeze stuff when selecting a bunch of lines at once.

Now the selection model is kept small, at the loss of future flexibility
to allow non-contiguous selections, which isn't supported in the current
version anyway.

Signed-off-by: flow <flowlnlnln@gmail.com>